// src/app/models/memoria.modelo.ts
export interface Memoria {
    id: string

    unoPalabra: string
    unoDescripcion: string

    dosPalabra: string
    dosDescripcion: string

    tresPalabra: string
    tresDescripcion: string

    cuatroPalabra: string
    cuatroDescripcion: string

    cincoPalabra: string
    cincoDescripcion: string

    seisPalabra: string
    seisDescripcion: string

    sietePalabra: string
    sieteDescripcion: string

    ochoPalabra: string
    ochoDescripcion: string

    nuevePalabra: string
    nueveDescripcion: string

    diezPalabra: string
    diezDescripcion: string

    oncePalabra: string
    onceDescripcion: string

    docePalabra: string
    doceDescripcion: string

    trecePalabra: string
    treceDescripcion: string

    catorcePalabra: string
    catorceDescripcion: string

    quincePalabra: string
    quinceDescripcion: string

    dieciseisPalabra: string
    dieciseisDescripcion: string

    diecisietePalabra: string
    diecisieteDescripcion: string

    dieciochoPalabra: string
    dieciochoDescripcion: string

    diecinuevePalabra: string
    diecinueveDescripcion: string

    veintePalabra: string
    veinteDescripcion: string

    updatedAt: number
    _deleted: boolean
}